# Context
592ec167ac598258dc986139b0cca7f84c219fa4 is very slow, because `git remote show` is executed each view rendering :tired_face:

So I add caching

ref #8

# Architecture
* Add default branch cache (in memory)
* Use cache from view rendering
* Update cache after each `git checkout`

# Benchmark at project page
Using Chrome developer tool

![image](https://user-images.githubusercontent.com/608755/58365124-b4b20480-7efa-11e9-963c-bd3c6b284705.png)

## Before (v1.1.4)
* 1,950ms
* 2,440ms
* 1,970ms
* 3,150ms
* 2,040ms

## After (this PR)
* 2,030ms (1st),  2ms (2nd)
* 1,960ms (1st),  5ms (2nd)
* 1,950ms (1st),  2ms (2nd)
* 2,010ms (1st),  2ms (2nd)
* 3,290ms (1st),  2ms (2nd)

### remarks
1st is slow because branch is not cached, but 2nd is fast because branch is cached.